Extract createDefaultLocalMavenRepository to LocalMavenRepository.kt
